The full FoBAAR model V5.2
This is a complete update to V5.1, including the following bug fixes:
identified by Min Lee - changes to the assignLAI.f90 code which resolve
an issue associated with PAR interception in early hours.
identified by Min Chen - changes in the way soil water is treated,
which allow soil water stress to be applied only after a certain
threshold is passed.

To do:
File paths are poorly coded in the current version. Update code to
allow paths to extend from current working directory.
Optimization parameters are hard coded. Change to allow this to be read
in from file.
Relic code: Soil single and two pool models should be removed, along
with relic inner and outer loop code.